+++ auth/src/main/java/com/sparta/msa_exam/auth/user/AuthController.java
@@ -1,8 +1,6 @@
 package com.sparta.msa_exam.auth.user;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
