+++ order/src/main/java/com/sparta/msa_exam/order/order/OrderController.java
@@ -1,8 +1,6 @@
 package com.sparta.msa_exam.order.order;
 
-import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
